Fixed memory leak issue during stream-parse
During stream-parse, the parser only deletes the previous element from
the node to free memory. However, this is not accounting for any
unwanted nodes between the targets. For example, if there is a new line
character between tags, those TextNodes did not get cleaned.

After each iteration, a new TextNode got added as previous siblings of
the next target, causing the performance to slow down with the number of
elements parsed.